refactor: rename ctx package to contextutil to avoid stdlib collision
- Rename internal/ctx to internal/contextutil
- Update package declarations in all affected files
- Update all import statements across codebase
- Update usage from ctx. to contextutil. in handler and tests

Resolves #87
