+++ internal/contextutil/context.go
@@ -1,4 +1,4 @@
-package ctx
+package contextutil
 
 import (
 	"fmt"
