+++ tests/context_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/vahiiiid/go-rest-api-boilerplate/internal/auth"
-	"github.com/vahiiiid/go-rest-api-boilerplate/internal/ctx"
+	"github.com/vahiiiid/go-rest-api-boilerplate/internal/contextutil"
 )
 
 func TestGetUser(t *testing.T) {
@@ -18,7 +18,7 @@ func TestGetUser(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		result := ctx.GetUser(c)
+		result := contextutil.GetUser(c)
 		assert.NotNil(t, result)
 		assert.Equal(t, uint(42), result.UserID)
 		assert.Equal(t, "test@example.com", result.Email)
@@ -27,7 +27,7 @@ func TestGetUser(t *testing.T) {
 	t.Run("returns nil when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		result := ctx.GetUser(c)
+		result := contextutil.GetUser(c)
 		assert.Nil(t, result)
 	})
 
@@ -35,7 +35,7 @@ func TestGetUser(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		result := ctx.GetUser(c)
+		result := contextutil.GetUser(c)
 		assert.Nil(t, result)
 	})
 }
@@ -48,7 +48,7 @@ func TestMustGetUser(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		result, err := ctx.MustGetUser(c)
+		result, err := contextutil.MustGetUser(c)
 		assert.NoError(t, err)
 		assert.NotNil(t, result)
 		assert.Equal(t, uint(42), result.UserID)
@@ -58,7 +58,7 @@ func TestMustGetUser(t *testing.T) {
 	t.Run("returns error when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		result, err := ctx.MustGetUser(c)
+		result, err := contextutil.MustGetUser(c)
 		assert.Error(t, err)
 		assert.Nil(t, result)
 		assert.Contains(t, err.Error(), "user not found in context")
@@ -73,14 +73,14 @@ func TestGetUserID(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		userID := ctx.GetUserID(c)
+		userID := contextutil.GetUserID(c)
 		assert.Equal(t, uint(42), userID)
 	})
 
 	t.Run("returns 0 when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		userID := ctx.GetUserID(c)
+		userID := contextutil.GetUserID(c)
 		assert.Equal(t, uint(0), userID)
 	})
 
@@ -88,7 +88,7 @@ func TestGetUserID(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		userID := ctx.GetUserID(c)
+		userID := contextutil.GetUserID(c)
 		assert.Equal(t, uint(0), userID)
 	})
 }
@@ -101,7 +101,7 @@ func TestMustGetUserID(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		userID, err := ctx.MustGetUserID(c)
+		userID, err := contextutil.MustGetUserID(c)
 		assert.NoError(t, err)
 		assert.Equal(t, uint(42), userID)
 	})
@@ -109,7 +109,7 @@ func TestMustGetUserID(t *testing.T) {
 	t.Run("returns error when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		userID, err := ctx.MustGetUserID(c)
+		userID, err := contextutil.MustGetUserID(c)
 		assert.Error(t, err)
 		assert.Equal(t, uint(0), userID)
 		assert.Contains(t, err.Error(), "user ID not found in context")
@@ -120,7 +120,7 @@ func TestMustGetUserID(t *testing.T) {
 		claims := &auth.Claims{UserID: 0, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		userID, err := ctx.MustGetUserID(c)
+		userID, err := contextutil.MustGetUserID(c)
 		assert.Error(t, err)
 		assert.Equal(t, uint(0), userID)
 		assert.Contains(t, err.Error(), "user ID not found in context")
@@ -135,14 +135,14 @@ func TestGetEmail(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		email := ctx.GetEmail(c)
+		email := contextutil.GetEmail(c)
 		assert.Equal(t, "test@example.com", email)
 	})
 
 	t.Run("returns empty string when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		email := ctx.GetEmail(c)
+		email := contextutil.GetEmail(c)
 		assert.Equal(t, "", email)
 	})
 
@@ -150,7 +150,7 @@ func TestGetEmail(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		email := ctx.GetEmail(c)
+		email := contextutil.GetEmail(c)
 		assert.Equal(t, "", email)
 	})
 }
@@ -163,14 +163,14 @@ func TestIsAuthenticated(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		authenticated := ctx.IsAuthenticated(c)
+		authenticated := contextutil.IsAuthenticated(c)
 		assert.True(t, authenticated)
 	})
 
 	t.Run("returns false when user is not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		authenticated := ctx.IsAuthenticated(c)
+		authenticated := contextutil.IsAuthenticated(c)
 		assert.False(t, authenticated)
 	})
 
@@ -178,7 +178,7 @@ func TestIsAuthenticated(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		authenticated := ctx.IsAuthenticated(c)
+		authenticated := contextutil.IsAuthenticated(c)
 		assert.False(t, authenticated)
 	})
 }
@@ -191,7 +191,7 @@ func TestCanAccessUser(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		canAccess := ctx.CanAccessUser(c, 42)
+		canAccess := contextutil.CanAccessUser(c, 42)
 		assert.True(t, canAccess)
 	})
 
@@ -200,14 +200,14 @@ func TestCanAccessUser(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com"}
 		c.Set(auth.KeyUser, claims)
 
-		canAccess := ctx.CanAccessUser(c, 43)
+		canAccess := contextutil.CanAccessUser(c, 43)
 		assert.False(t, canAccess)
 	})
 
 	t.Run("returns false when user is not authenticated", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		canAccess := ctx.CanAccessUser(c, 42)
+		canAccess := contextutil.CanAccessUser(c, 42)
 		assert.False(t, canAccess)
 	})
 }
@@ -220,14 +220,14 @@ func TestGetUserName(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com", Name: "John Doe"}
 		c.Set(auth.KeyUser, claims)
 
-		userName := ctx.GetUserName(c)
+		userName := contextutil.GetUserName(c)
 		assert.Equal(t, "John Doe", userName)
 	})
 
 	t.Run("returns empty string when not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		userName := ctx.GetUserName(c)
+		userName := contextutil.GetUserName(c)
 		assert.Equal(t, "", userName)
 	})
 
@@ -235,7 +235,7 @@ func TestGetUserName(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		userName := ctx.GetUserName(c)
+		userName := contextutil.GetUserName(c)
 		assert.Equal(t, "", userName)
 	})
 }
@@ -248,14 +248,14 @@ func TestHasRole(t *testing.T) {
 		claims := &auth.Claims{UserID: 42, Email: "test@example.com", Name: "John Doe"}
 		c.Set(auth.KeyUser, claims)
 
-		hasRole := ctx.HasRole(c, "admin")
+		hasRole := contextutil.HasRole(c, "admin")
 		assert.False(t, hasRole) // Currently returns false as roles are not implemented
 	})
 
 	t.Run("returns false when user is not present", func(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 
-		hasRole := ctx.HasRole(c, "admin")
+		hasRole := contextutil.HasRole(c, "admin")
 		assert.False(t, hasRole)
 	})
 
@@ -263,7 +263,7 @@ func TestHasRole(t *testing.T) {
 		c, _ := gin.CreateTestContext(nil)
 		c.Set(auth.KeyUser, "not-a-claims-struct")
 
-		hasRole := ctx.HasRole(c, "admin")
+		hasRole := contextutil.HasRole(c, "admin")
 		assert.False(t, hasRole)
 	})
 }
